tests/wpt/web-platform-tests/html/browsers/windows/nested-browsing-contexts/frameElement.sub.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<title>HTML Test: window.frameElement</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

var t1 = async_test("The window's frameElement attribute must return its container element if it is a nested browsing context");
var t2 = async_test("The SecurityError must be thrown if the container's document does not have the same effective script origin");

function on_load() {
  t1.step(function () {
    assert_equals(frames[0].frameElement, document.getElementById("fr1"),
                  "The frameElement attribute should be the first iframe element.");
    assert_equals(window["win2"].frameElement, document.getElementById("obj"),
                  "The frameElement attribute should be the object element.");
    assert_equals(window["win3"].frameElement, document.getElementById("emb"),
                  "The frameElement attribute should be the embed element.");
    assert_equals(document.getElementById("fr3").contentWindow[0].frameElement,
                  document.getElementById("fr3").contentDocument.getElementById("f1"),
                 "The frameElement attribute should be the frame element in 'test.html'.");
  });
  t1.done();

  t2.step(function () {
    assert_throws("SecurityError", function () { frames[1].frameElement; },
                  "The SecurityError exception should be thrown.");
  });
  t2.done();
}

</script>
<body onload="on_load()">
  <div id="log"></div>
  <iframe id="fr1"></iframe>
  <iframe id="fr2" src="test.html"></iframe>
  <iframe id="fr3" src="" style="display:none"></iframe>
  <object id="obj" name="win2" type="text/html" data="about:blank"></object>
  <embed id="emb" name="win3" type="image/svg+xml" src="/images/green.svg" />
  <script>

  setup(function () {
    var src = "http://{{domains[www1]}}:{{ports[http][0]}}";
    src += document.location.pathname.substring(0, document.location.pathname.lastIndexOf("/") + 1);
    src += "test.html";
    document.getElementById("fr2").src = src;
  });

  test(function () {
    assert_equals(window.frameElement, null,
                  "The frameElement attribute should be null.");
  }, "The window's frameElement attribute must return null if it is not a nested browsing context");

  </script>
</body>
